refactor: navigation components
